[RPC] Add Streaming RPC Types
Closes: #8593
