+++ json-rpc/src/stream_rpc/mod.rs
@@ -18,7 +18,6 @@ mod logging;
 pub mod startup;
 
 mod connection;
-mod errors;
 mod json_rpc;
 mod subscription_types;
 mod subscriptions;
