+++ json-rpc/types/src/lib.rs
@@ -1,6 +1,8 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod stream;
+
 pub mod errors;
 pub mod request;
 pub mod response;
