+++ json-rpc/types/src/stream/errors.rs
