+++ sdk/client/Cargo.toml
@@ -23,7 +23,7 @@ serde = { version = "1.0.124", features = ["derive"] }
 serde_json = "1.0.64"
 
 diem-crypto = { path = "../../crypto/crypto", version = "0.0.2" }
-diem-json-rpc-types  = { path = "../../json-rpc/types", version = "0.0.2" }
+diem-json-rpc-types  = { path = "../../json-rpc/types", version = "0.0.3" }
 move-core-types = { path = "../../language/move-core/types", version = "0.0.2" }
 diem-types = { path = "../../types", version = "0.0.2" }
